Rename `classEquityChart.py` -> `equity_chart.py`.
