+++ src/gui/equity_chart.py
